Split into test and env code
